Convert embedded quotes in strings correctly
The regular expression which identifies strings literals in source
can not handle the case when a quote character is embedded in the
string. For example "Convert \"this\"" would not convert correctly
due to the embedded \" characters in the string. In order to avoid
this situation we can preprocess all the escape sequences in the source
prior to trying to identify string and character literals.